[TASK] Clean up tests for root Index.rst generation test
