+++ tests/Functional/Typo3BackendViewhelperRootIndexRstDocumentationGenerationTest.php
@@ -1,0 +1,63 @@
+<?php
+declare(strict_types=1);
+
+namespace NamelessCoder\FluidDocumentationGenerator\Tests\Functional;
+
+use NamelessCoder\FluidDocumentationGenerator\Data\DataFileResolver;
+use NamelessCoder\FluidDocumentationGenerator\Export\RstExporter;
+use NamelessCoder\FluidDocumentationGenerator\SchemaDocumentationGenerator;
+use org\bovigo\vfs\vfsStream;
+use org\bovigo\vfs\vfsStreamDirectory;
+use PHPUnit\Framework\TestCase;
+
+class Typo3BackendViewhelperRootIndexRstDocumentationGenerationTest extends TestCase
+{
+    /**
+     * @var vfsStreamDirectory
+     */
+    private $vfs;
+
+    /**
+     * the generated file is compared against this fixture file
+     * @var string
+     */
+    private $fixtureFilePath = __DIR__ . '/../Fixtures/output/Documentation/Index.rst';
+
+    /**
+     * output of the generation process
+     * @var string
+     */
+    private $generatedFilePath = 'outputDir/public/Index.rst';
+
+    protected function setUp()
+    {
+        $this->vfs = vfsStream::setup('outputDir');
+        $this->vfs->addChild(vfsStream::newDirectory('cache'));
+        $dataFileResolver = DataFileResolver::getInstance(vfsStream::url('outputDir'));
+        $dataFileResolver->setResourcesDirectory(__DIR__ . '/../../resources/');
+        $dataFileResolver->setSchemasDirectory(__DIR__ . '/../Fixtures/schemas/typo3');
+        $schemaDocumentationGenerator = new SchemaDocumentationGenerator(
+            [
+                new RstExporter()
+            ]
+        );
+        $schemaDocumentationGenerator->generateFilesForRoot();
+    }
+
+    /**
+     * @test
+     */
+    public function fileIsCreated()
+    {
+        $this->assertTrue($this->vfs->hasChild($this->generatedFilePath));
+    }
+
+    /**
+     * @test
+     */
+    public function includeClausePointsToSettingsCfg()
+    {
+        $output = file($this->vfs->getChild($this->generatedFilePath)->url());
+        $this->assertSame('.. include:: Includes.txt' . PHP_EOL, $output[0]);
+    }
+}
